move loading/writing of events and onsets
